HB-2 add monday getter function
i added a modal for the pending requests that prompts the user for a reason to approve/deny the request. next steps are to work with daksh to setup a push to the api when this happens
the active requests card is now able to display without the andrew wang step.
the "create booking" date selector now shows the entire work week rather than one day. next steps are to make it phone-friendly
increased the touch targets for the "reason" selector
requests no longer have a GUID attached to them since i am not sure if we'll be giving each request one. originally the purpose of attaching a guid to each request is for tracking, but thats no longer needed. will have to discuss with team whether to readd it