Made schedule selector functional
Schedule selector can now be used to choose a time range and date. Both time range and date are displayed on booking confirmation screen. User is prevented from finishing their booking until a valid time range and date are selected. Made it so that user can only select a time range for a single day.